+++ assets/yii2-dynamic-form.js
@@ -7,6 +7,11 @@
  */
 (function ($) {
     var pluginName = 'yiiDynamicForm';
+
+    var regexID = /^(.+?)([-\d-]{1,})(.+)$/i;
+
+    var regexName = /(^.+?)([\[\d{1,}\]]{1,})(\[.+\]$)/i;
+
     $.fn.yiiDynamicForm = function (method) {
         if (methods[method]) {
             return methods[method].apply(this, Array.prototype.slice.call(arguments, 1));
@@ -19,169 +24,323 @@
     };
 
     var events = {
-        beforeClone: 'beforeClone',
-        afterClone: 'afterClone',
-        limitReached: 'limitReached',
-        afterRemove: 'afterRemove'
+        beforeInsert: 'beforeInsert',
+        afterInsert: 'afterInsert',
+        beforeDelete: 'beforeDelete',
+        afterDelete: 'afterDelete',
+        limitReached: 'limitReached'
     };
-
-    var defaults = {
-        // the maximum times, an element can be cloned
-        limit: 999, //setting it to a high number, by default
-        min: 1,
-        dynamicItems: '.dynamic-items',
-        dynamicItem: '.dynamic-item',
-        cloneButton: '.clone',
-        deleteButton: '.delete',
-        clonePosition: 'after', // or 'before'
-        template: false
-    };
-
-    var regex = /(.*-)(\d)(-.*)/i;
 
     var methods = {
-        init: function (options) {
+        init: function (settings) {
             return this.each(function () {
-                var $container = $(this);
+                var $this = $(this);
+                var widgetData = {};
 
-                var data = $container.data(pluginName);
-                if (!data) {
-                    var settings = $.extend({}, defaults, options || {});
+                if ($('body').data(pluginName) !== undefined) {
+                    widgetData = $('body').data(pluginName);
+                }
 
-                    settings.fields.forEach(function(attribute) {
-                        var yiiActiveFormAttribute = $("#" + settings.formId).yiiActiveForm("find", attribute.id.replace("{}", 0));
-                        if (yiiActiveFormAttribute !== undefined) {
-                           attribute.baseConfig = yiiActiveFormAttribute;
-                        }
+                if (widgetData[settings.widgetContainer] === undefined) {
+                    widgetData = (widgetData[settings.widgetContainer] = settings, widgetData);
+                    settings.template = _parseTemplate(settings);
+
+                    $('body').data(pluginName, widgetData);
+
+                    // add a click handler for the clone button
+                    $('#' + settings.formId).on('click', settings.insertButton, function(e) {
+                        e.preventDefault();
+                        $('.' + settings.widgetContainer).triggerHandler(events.beforeInsert, [$(this)]);
+                        _addItem(settings.widgetContainer, e, $(this));
                     });
 
-                    $container.data(pluginName, {
-                        target : $container,
-                        settings: settings
+                    // add a click handler for the remove button
+                    $('#' + settings.formId).on('click', settings.deleteButton, function(e) {
+                        e.preventDefault();
+                        _deleteItem(settings.widgetContainer, e, $(this));
                     });
-
-                    if ($container.find(settings.cloneButton).length) {
-                        // add a click handler for the clone button
-                        $container.on('click', settings.cloneButton, function(event) {
-                            event.preventDefault();
-                            $container.triggerHandler(events.beforeClone, [$(this)]);
-                            startClone(event, $(this), $container);
-                        });
-                    }
-
-                    $container.on('click', settings.deleteButton, function(event) {
-                        event.preventDefault();
-                        removeItem($container, $(this));
-                    });
-
-                    if ($container.find(settings.dynamicItem).length === 0) {
-                        settings.fields.forEach(function(attribute) {
-                            var id = attribute.id.replace("{}", 0);
-                            if ($("#" + settings.formId).yiiActiveForm("find", id) !== undefined) {
-                               $("#" + settings.formId).yiiActiveForm("remove", id);
-                            }
-                        });
-                    }
                 }
             });
         },
 
         updateContainer: function () {
-            var $container = $(this);
-            redoIDs($container);
-            restoreSpecialJs($container);
-            fixFormValidaton($container);
+            var widgetContainer = $(this).attr('data-dynamicform');
+            _updateAttributes(widgetContainer);
+            _restoreSpecialJs(widgetContainer);
+            _fixFormValidaton(widgetContainer);
         }
     };
 
-    var startClone = function(event, $btnAdd, $container) {
-        var config = $container.data(pluginName).settings;
-
-        // get the count of all the clones
-        var cloneCount = $container.find(config.dynamicItem).length;
-        // check if we've reached the maximum limit
-        if (cloneCount < config.limit) {
-            $toclone = $($(config.template).html());
-            // clone it
-            $newclone = $toclone.clone(false, false);
-            $newclone.find('input, textarea, select').each(function() {
-                $(this).val('');
-            });
-            if (config.clonePosition != 'after') {
-                $(config.dynamicItems).prepend($newclone);
-            } else {
-                $(config.dynamicItems).append($newclone);
+    var _parseTemplate = function(settings) {
+        var $template = $(settings.template);
+        $template.find('div[data-dynamicform]').each(function(){
+            var widgetData = $('body').data(pluginName)[$(this).attr('data-dynamicform')];
+            if ($(widgetData.widgetItem).length > 1) {
+                var item = $(this).find(widgetData.widgetItem).first()[0].outerHTML;
+                $(this).find(widgetData.widgetBody).html(item);
             }
-            redoIDs($container);
-            removeErrorCssClass($newclone);
-            restoreSpecialJs($container);
-            fixFormValidaton($container);
-            $container.triggerHandler(events.afterClone, $newclone);
+        });
+
+        $template.find('input, textarea, select').each(function() {
+            $(this).val('');
+        });
+
+        $template.find('input[type="checkbox"]').each(function() {
+            var inputName = $(this).attr('name');
+            var $inputHidden = $template.find('input[type="hidden"][name="' + inputName + '"]').first();
+            if ($inputHidden) {
+                $(this).val(1);
+                $inputHidden.val(0);
+            }
+        });
+
+        return $template;
+    };
+
+    var _getDataWidgetRoot = function(widgetContainer) {
+        var widgetData = $('body').data(pluginName)[widgetContainer];
+        var widgetContainerRoot = $(widgetData.widgetBody).parents('div[data-dynamicform]').last().attr('data-dynamicform');
+        return $('body').data(pluginName)[widgetContainerRoot];
+    };
+
+    var _getLevel = function($elem) {
+        var level = $elem.parents('div[data-dynamicform]').length;
+        level = (level < 0) ? 0 : level;
+        return level;
+    };
+
+    var _count = function($elem, widgetContainer) {
+        var widgetData = $('body').data(pluginName)[widgetContainer];
+        return $elem.closest('.' + widgetContainer).find(widgetData.widgetItem).length;
+    };
+
+    var _creatIdentifiers = function(level) {
+        return new Array(level + 2).join('0').split('');
+    };
+
+    var _addItem = function(widgetContainer, e, $elem) {
+        var widgetData = $('body').data(pluginName)[widgetContainer];
+        var count = _count($elem, widgetContainer);
+
+        if (count < widgetData.limit) {
+            $toclone = widgetData.template;
+            $newclone = $toclone.clone(false, false);
+
+            if (widgetData.insertPosition === 'top') {
+                $elem.closest('.' + widgetData.widgetContainer).find(widgetData.widgetBody).prepend($newclone);
+            } else {
+                $elem.closest('.' + widgetData.widgetContainer).find(widgetData.widgetBody).append($newclone);
+            }
+
+            _updateAttributes(widgetContainer);
+            _restoreSpecialJs(widgetContainer);
+            _fixFormValidaton(widgetContainer);
+            $elem.closest('.' + widgetContainer).triggerHandler(events.afterInsert, $newclone);
         } else {
             // trigger a custom event for hooking
-            $container.triggerHandler(events.limitReached, config.limit);
+            $elem.closest('.' + widgetContainer).triggerHandler(events.limitReached, widgetData.limit);
         }
     };
 
-    var fixFormValidaton = function($container) {
-        var config = $container.data(pluginName).settings;
-        $(config.dynamicItem).each(function(i) {
-            config.fields.forEach(function(v) {
-                var id = v.id.replace("{}", i);
-                var attribute = v.baseConfig;
-                if (attribute !== undefined) {
-                    attribute = $.extend(true, {}, attribute);
-                    attribute.id = id;
-                    attribute.container = ".field-" + id;
-                    attribute.input = "#" + id;
-                    attribute.name = v.name.replace("{}", i);
-                    attribute.value = $("#" + id).val();
-                    attribute.status = 0;
-                    if ($("#" + config.formId).yiiActiveForm("find", id) !== "undefined") {
-                        $("#" + config.formId).yiiActiveForm("remove", id);
-                    }
-                    $("#" + config.formId).yiiActiveForm("add", attribute);
+    var _removeValidations = function($elem, widgetData, count) {
+        if (count > 1) {
+            $elem.find('div[data-dynamicform]').each(function() {
+                var currentAttrData = $('body').data(pluginName)[$(this).attr('data-dynamicform')];
+                var level           = _getLevel($(this));
+                var identifiers     = _creatIdentifiers(level);
+                var numItems        = $(this).find(currentAttrData.widgetItem).length;
+
+                for (var i = 1; i <= numItems -1; i++) {
+                    var aux = identifiers;
+                    aux[level] = i;
+                    currentAttrData.fields.forEach(function(input) {
+                        var id = input.id.replace("{}", aux.join('-'));
+                        if ($("#" + currentAttrData.formId).yiiActiveForm("find", id) !== "undefined") {
+                            $("#" + currentAttrData.formId).yiiActiveForm("remove", id);
+                        }
+                    });
+                }
+            });
+
+            var level          = _getLevel($elem.closest('.' + widgetData.widgetContainer));
+            var rootData       = _getDataWidgetRoot(widgetData.widgetContainer);
+            var identifiers    = _creatIdentifiers(level);
+            identifiers[0]     = $(rootData.widgetItem).length - 1;
+            identifiers[level] = count - 1;
+
+            widgetData.fields.forEach(function(input) {
+                var id = input.id.replace("{}", identifiers.join('-'));
+                if ($("#" + widgetData.formId).yiiActiveForm("find", id) !== "undefined") {
+                    $("#" + widgetData.formId).yiiActiveForm("remove", id);
+                }
+            });
+        }
+    };
+
+    var _deleteItem = function(widgetContainer, e, $elem) {
+        var widgetData = $('body').data(pluginName)[widgetContainer];
+        var count = _count($elem, widgetContainer);
+
+        if (count > widgetData.min) {
+            $todelete = $elem.closest(widgetData.widgetItem);
+
+            // trigger a custom event for hooking
+            var eventResult = $('.' + widgetData.widgetContainer).triggerHandler(events.beforeDelete, $todelete);
+            if (eventResult !== false) {
+                _removeValidations($todelete, widgetData, count);
+                $todelete.remove();
+                _updateAttributes(widgetContainer);
+                _restoreSpecialJs(widgetContainer);
+                _fixFormValidaton(widgetContainer);
+                $('.' + widgetData.widgetContainer).triggerHandler(events.afterDelete);
+            }
+        }
+    };
+
+    var _updateAttrID = function($item, $elem, index, currentAttrData, level) {
+        var id = $elem.attr('id');
+        var newID = id;
+
+        if (id !== undefined) {
+            var matches = id.match(regexID);
+
+            if (matches && matches.length === 4) {
+                matches[2] = matches[2].substring(1, matches[2].length - 1);
+                var identifiers = matches[2].split('-');
+                identifiers[0] = index;
+
+                if (level > 0 && identifiers[level] !== undefined) {
+                    identifiers[level] = $elem.closest(currentAttrData.widgetItem).index();
+                }
+
+                newID = matches[1] + '-' + identifiers.join('-') + '-' + matches[3];
+                $elem.attr('id', newID);
+            } else {
+                newID = id + index;
+                $elem.attr('id', newID);
+            }
+        }
+
+        if (id !== newID) {
+            $item.find('.field-' + id).each(function() {
+                $(this).removeClass('field-' + id).addClass('field-' + newID);
+            });
+            // update "for" attribute
+            $elem.closest($item).find("label[for='" + id + "']").attr('for',newID); 
+        }
+
+        return newID;
+    };
+
+    var _updateAttrName = function($item, $elem, index, currentAttrData, level) {
+        var name = $elem.attr('name');
+
+        if (name !== undefined) {
+            var matches = name.match(regexName);
+
+            if (matches && matches.length === 4) {
+                matches[2] = matches[2].replace(/\]\[/g, "-").replace(/\]|\[/g, '');
+                var identifiers = matches[2].split('-');
+                identifiers[0] = index;
+
+                if (level > 0 && identifiers[level] !== undefined) {
+                    identifiers[level] = $elem.closest(currentAttrData.widgetItem).index();
+                }
+
+                name = matches[1] + '[' + identifiers.join('][') + ']' + matches[3];
+                $elem.attr('name', name);
+            }
+        }
+
+        return name;
+    };
+
+    var _updateAttributes = function(widgetContainer) {
+        var rootData = _getDataWidgetRoot(widgetContainer);
+
+        $(rootData.widgetItem).each(function(index) {
+            var $item = $(this);
+            var level = 0;
+            var currentAttrData = rootData;
+
+            $(this).find('*').each(function() {
+                if ($(this).attr('data-dynamicform') !== undefined) {
+                    currentAttrData = $('body').data(pluginName)[$(this).attr('data-dynamicform')];
+                    level = _getLevel($(this));
+                }
+
+                // update "id" attribute
+                _updateAttrID($item, $(this), index, currentAttrData, level);
+
+                // update "name" attribute
+                _updateAttrName($item, $(this), index, currentAttrData, level);
+            });
+        });
+    };
+
+    var _fixFormValidatonInput = function(widgetData, input, index, id, name) {
+        var attribute = input.baseConfig;
+
+        if (attribute !== undefined) {
+            attribute = $.extend(true, {}, attribute);
+            attribute.id = id;
+            attribute.container = ".field-" + id;
+            attribute.input = "#" + id;
+            attribute.name = input.name.replace("{}", index);
+            attribute.value = $("#" + id).val();
+            attribute.status = 0;
+
+            if ($("#" + widgetData.formId).yiiActiveForm("find", id) !== "undefined") {
+                $("#" + widgetData.formId).yiiActiveForm("remove", id);
+            }
+
+            $("#" + widgetData.formId).yiiActiveForm("add", attribute);
+        }
+    };
+
+    var _fixFormValidatonInputs = function(widgetData, level, index, i) {
+         widgetData.fields.forEach(function(input) {
+            var identifiers    = _creatIdentifiers(level);
+            var yiiActiveFormAttribute = $("#" + widgetData.formId).yiiActiveForm("find", input.id.replace("{}", identifiers.join('-')));
+
+            if (yiiActiveFormAttribute !== undefined) {
+                input.baseConfig       = yiiActiveFormAttribute;
+                identifiers[0]     = index;
+                identifiers[level] = i;
+                var id             = input.id.replace("{}", identifiers.join('-'));
+                var name           = input.name.replace("{}", identifiers.join(']['));
+                _fixFormValidatonInput(widgetData, input, i, id, name);
+            }
+        });
+    };
+
+    var _fixFormValidaton = function(widgetContainer) {
+        var rootData = _getDataWidgetRoot(widgetContainer);
+
+        $(rootData.widgetItem).each(function(index) {
+            var $item = $(this);
+            var level = 0;
+            var currentAttrData = rootData;
+            _fixFormValidatonInputs(rootData, level, index, index);
+
+            $(this).find('div[data-dynamicform]').each(function() {
+                currentAttrData = $('body').data(pluginName)[$(this).attr('data-dynamicform')];
+                level = _getLevel($(this));
+
+                if (level > 0) {
+                    $(this).find(currentAttrData.widgetItem).each(function(i) {
+                         _fixFormValidatonInputs(currentAttrData, level, index, i);
+                    });
                 }
             });
         });
     };
 
-    var removeItem = function($container, $item) {
-        var config = $container.data(pluginName).settings;
-        var cloneCount = $container.find(config.dynamicItem).length;
-        if (cloneCount > config.min) {
-            // get the closest parent clone
-            $todelete = $item.closest(config.dynamicItem);
-            $todelete.remove();
-            redoIDs($container);
-            restoreSpecialJs($container);
-            fixFormValidaton($container);
-            if (cloneCount > 0) {
-                config.fields.forEach(function(v) {
-                    var id = v.id.replace("{}", (cloneCount - 1));
-                    if ($("#" + config.formId).yiiActiveForm("find", id) !== "undefined") {
-                        $("#" + config.formId).yiiActiveForm("remove", id);
-                    }
-                });
-            }
-            $container.triggerHandler(events.afterRemove);
-        }
-    }
-
-    var removeErrorCssClass = function($container) {
-        $container.find('.has-success').each(function() {
-            $(this).removeClass('has-success');
-        });
-        $container.find('.has-error').each(function() {
-            $(this).removeClass('has-error');
-        });
-    };
-
-    var restoreSpecialJs = function($container) {
-        var config = $container.data(pluginName).settings;
+    var _restoreSpecialJs = function(widgetContainer) {
+        var rootData = _getDataWidgetRoot(widgetContainer);
 
         // "kartik-v/yii2-widget-datepicker"
-        var $hasDatepicker = $(config.dynamicItems).find('[data-krajee-datepicker]');
+        var $hasDatepicker = $(rootData.widgetItem).find('[data-krajee-datepicker]');
         if ($hasDatepicker.length > 0) {
             $hasDatepicker.each(function() {
                 $(this).parent().removeData().datepicker('remove');
@@ -190,7 +349,7 @@
         }
 
         // "kartik-v/yii2-widget-timepicker"
-        var $hasTimepicker = $(config.dynamicItems).find('[data-krajee-timepicker]');
+        var $hasTimepicker = $(rootData.widgetItem).find('[data-krajee-timepicker]');
         if ($hasTimepicker.length > 0) {
             $hasTimepicker.each(function() {
                 $(this).removeData().off();
@@ -201,7 +360,7 @@
         }
 
         // "kartik-v/yii2-money"
-        var $hasMaskmoney = $(config.dynamicItems).find('[data-krajee-maskMoney]');
+        var $hasMaskmoney = $(rootData.widgetItem).find('[data-krajee-maskMoney]');
         if ($hasMaskmoney.length > 0) {
             $hasMaskmoney.each(function() {
                 $(this).parent().find('input').removeData().off();
@@ -219,7 +378,7 @@
         }
 
         // "kartik-v/yii2-widget-fileinput"
-        var $hasFileinput = $(config.dynamicItems).find('[data-krajee-fileinput]');
+        var $hasFileinput = $(rootData.widgetItem).find('[data-krajee-fileinput]');
         if ($hasFileinput.length > 0) {
             $hasFileinput.each(function() {
                 $(this).fileinput(eval($(this).attr('data-krajee-fileinput')));
@@ -227,7 +386,7 @@
         }
 
         // "kartik-v/yii2-widget-touchspin"
-        var $hasTouchSpin = $(config.dynamicItems).find('[data-krajee-TouchSpin]');
+        var $hasTouchSpin = $(rootData.widgetItem).find('[data-krajee-TouchSpin]');
         if ($hasTouchSpin.length > 0) {
             $hasTouchSpin.each(function() {
                 $(this).TouchSpin('destroy');
@@ -236,7 +395,7 @@
         }
 
         // "kartik-v/yii2-widget-colorinput"
-        var $hasSpectrum = $(config.dynamicItems).find('[data-krajee-spectrum]');
+        var $hasSpectrum = $(rootData.widgetItem).find('[data-krajee-spectrum]');
         if ($hasSpectrum.length > 0) {
             $hasSpectrum.each(function() {
                 var id = '#' + $(this).attr('id');
@@ -258,7 +417,7 @@
         }
 
         // "kartik-v/yii2-widget-depdrop"
-        var $hasDepdrop = $(config.dynamicItems).find('[data-krajee-depdrop]');
+        var $hasDepdrop = $(rootData.widgetItem).find('[data-krajee-depdrop]');
         if ($hasDepdrop.length > 0) {
             $hasDepdrop.each(function() {
                 $(this).removeData().off();
@@ -279,7 +438,7 @@
         }
 
         // "kartik-v/yii2-widget-select2"
-        var $hasSelect2 = $(config.dynamicItems).find('[data-krajee-select2]');
+        var $hasSelect2 = $(rootData.widgetItem).find('[data-krajee-select2]');
         if ($hasSelect2.length > 0) {
             $hasSelect2.each(function() {
                 var id = $(this).attr('id');
@@ -301,57 +460,6 @@
                 }
             });
         }
-    };
-
-    var redoIDs = function($container) {
-        var config = $container.data(pluginName).settings;
-        $container.find(config.dynamicItem).each(function(i) {
-            // first modify the clone id
-            var j = i;
-            var $obj = $(this);
-            $(this).find('*').each(function() {
-                var id = $(this).attr('id');
-                if (id) {
-                    var newID = id;
-                    var match = id.match(regex);
-                    if (match && match.length === 4) {
-                        newID = match[1] + i + match[3];
-                        $(this).attr('id', newID);
-                    } else {
-                        newID = id + i;
-                        $(this).attr('id', newID);
-                    }
-                    if (id !== newID) {
-                        $obj.find('.field-' + id).each(function() {
-                            $(this).removeClass('field-' + id).addClass('field-' + newID);
-                        });
-                    }
-                }
-                // "for" attribute 
-                var elmLabel = $(this).attr('for');
-                if (elmLabel) {
-                    var match = elmLabel.match(regex);
-                    if (match && match.length === 4) {
-                        $(this).attr('for', match[1] + i + match[3]);
-                    } else {
-                        $(this).attr('for', elmLabel + i);
-                    }
-                }
-                // "name" attribute 
-                var name = $(this).attr('name');
-                // This will increment the numeric array index for cloned field names
-                if (name) {
-                    var matches = name.match(/\[([^}]+)\]/);
-                    if (matches && matches.length >= 1) {
-                        var st = name;
-                        name = [].map.call(st, function(s, i) {
-                            return (!isNaN(+s) && st[i - 1] === '[' && st[i + 1] === ']') ? j : s;
-                        }).join('');
-                        $(this).attr('name', name);
-                    }
-                }
-            });
-        });
     };
 
 })(window.jQuery);